PYIC-7555 TS dev routes
